substrate_cli/requirements.txt
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
chromadb>=0.5.0
